requirements.txt
jupyter==1.0.0
numpy==1.21.2
pandas==1.3.2
pillow==8.3.2
albumentations==1.0.3
tensorflow==2.6.0
opencv-python-headless==4.5.3.56
segmentation-models==1.0.1
scikit-learn==0.24.2
matplotlib==3.4.3
